kmscube: gears: use only GLES2
The gears mode currently depends on either GLES3 or desktop GL to build,
but in both cases without a proper meson dependency check.
It also doesn't really use any features which would require those GL
versions over GLES2, which is already a global dependency of kmscube.
So just change it to use symbols which are available in GLES2 and it
can be built and run in all platforms.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>